error changes

+++ src/main/java/com/example/movie_reservation/serviceImplementation/ScreenServiceImpl.java
@@ -1,9 +1,12 @@
 package com.example.movie_reservation.serviceImplementation;
 
+import com.example.movie_reservation.exception.ResourceNotFoundException;
 import com.example.movie_reservation.model.Movie;
+import com.example.movie_reservation.model.Role;
 import com.example.movie_reservation.model.Screen;
 import com.example.movie_reservation.model.Theatre;
 import com.example.movie_reservation.repository.ScreenRepository;
+import com.example.movie_reservation.repository.SeatRepository;
 import com.example.movie_reservation.repository.TheatreRepository;
 import com.example.movie_reservation.requestDTO.ScreenRequestDTO;
 import com.example.movie_reservation.responseDTO.MovieResponseDTO;
@@ -13,6 +16,7 @@ import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
+import java.util.Optional;
 
 @Service
 @RequiredArgsConstructor
@@ -20,26 +24,27 @@ public class ScreenServiceImpl implements ScreenService {
 
     private final ScreenRepository screenRepository;
     private final TheatreRepository theatreRepository;
+    private final SeatRepository seatRepository;
 
     @Override
-    public ScreenResponseDTO createScreen(ScreenRequestDTO screen) {
+    public ScreenResponseDTO createScreen(ScreenRequestDTO screen) throws ResourceNotFoundException {
 
         //  Validating Theatre Exists or not
         Long theatreId = screen.getTheatre_id();
         Theatre theatre = theatreRepository.findById(theatreId)
-                .orElseThrow(() -> new RuntimeException("Theatre not found"));
+                .orElseThrow(() -> new ResourceNotFoundException("Theatre not found"));
 
         //  Validating Unique Screen Number per Theatre
         boolean exists = screenRepository.existsByScreenNameAndTheatre_TheatreId(
-                screen.getScreenName(), theatreId
-        );
+                screen.getScreenName(), theatreId);
+
         if (exists) {
-            throw new RuntimeException("Screen number already exists for this theatre");
+            throw new ResourceNotFoundException("Screen name already exists for this theatre");
         }
 
         //  Seat Count validation
         if (screen.getTotalSeatCount() <= 0) {
-            throw new RuntimeException("Total seat count must be greater than zero");
+            throw new ResourceNotFoundException("Total seat count must be greater than zero");
         }
 
         Screen screenRequest = Screen.builder()
@@ -54,37 +59,61 @@ public class ScreenServiceImpl implements ScreenService {
     }
 
     @Override
-    public Screen getScreenById(Long screenId) {
-        return screenRepository.findById(screenId)
-                .orElseThrow(() -> new RuntimeException("Screen not found"));
+    public ScreenResponseDTO getScreenById(Long screenId) {
+        Screen screen = screenRepository.findById(screenId)
+                .orElseThrow(() -> new ResourceNotFoundException("Screen not found"));
+
+        return  mapToResponse(screen);
     }
 
     @Override
-    public List<Screen> getAllScreens() {
-        return screenRepository.findAll();
+    public List<ScreenResponseDTO> getAllScreens() {
+
+        List<Screen> screens = screenRepository.findAll();
+
+        return screens.stream()
+                .map(this::mapToResponse)
+                .toList();
+
     }
 
     @Override
-    public List<Screen> getScreensByTheatre(Long theatreId) {
+    public List<ScreenResponseDTO> getScreensByTheatre(Long theatreId) {
 
         // Checking Theatre Exists or not
         if (!theatreRepository.existsById(theatreId)) {
-            throw new RuntimeException("Theatre not found");
+            throw new ResourceNotFoundException("Theatre not found");
         }
 
-        return screenRepository.findByTheatre_TheatreId(theatreId);
+        //screenRepository.findById(theatreId)
+               // .orElseThrow(() ->
+                   //     new ResourceNotFoundException("Theatre not found: " + theatreId));
+
+        List<Screen> screens = screenRepository.findByTheatre_TheatreId(theatreId);
+
+        if (screens.isEmpty()) {
+            throw new ResourceNotFoundException("No screens found for theatre: " + theatreId);
+        }
+
+
+        return screens.stream()
+                .map(this::mapToResponse)
+                .toList();
+
+
     }
 
     @Override
     public ScreenResponseDTO updateScreen(Long screenId, ScreenRequestDTO screen) {
 
-        Screen existing = getScreenById(screenId);
+        Screen existing = screenRepository.findById(screenId)
+                .orElseThrow(() -> new ResourceNotFoundException("Screen not found"));
 
         Long theatreId = screen.getTheatre_id();
 
         // Checking Theatre Exists or not
         Theatre theatre = theatreRepository.findById(theatreId)
-                .orElseThrow(() -> new RuntimeException("Theatre not found"));
+                .orElseThrow(() -> new ResourceNotFoundException("Theatre not found"));
 
         // Validate screen name uniqueness (if changed)
         boolean exists = screenRepository.existsByScreenNameAndTheatre_TheatreId(
@@ -92,11 +121,11 @@ public class ScreenServiceImpl implements ScreenService {
         );
 
         if (exists && !existing.getScreenName().equals(screen.getScreenName())) {
-            throw new RuntimeException("Screen Name already exists for this theatre");
+            throw new ResourceNotFoundException("Screen Name already exists for this theatre");
         }
 
         if (screen.getTotalSeatCount() <= 0) {
-            throw new RuntimeException("Total seat count must be greater than zero");
+            throw new ResourceNotFoundException("Total seat count must be greater than zero");
         }
 
         existing.setScreenName(screen.getScreenName());
@@ -110,7 +139,14 @@ public class ScreenServiceImpl implements ScreenService {
 
     @Override
     public void deleteScreen(Long screenId) {
-        Screen screen = getScreenById(screenId);
+        Screen screen = screenRepository.findById(screenId)
+                .orElseThrow(() -> new ResourceNotFoundException("Screen not found"));
+
+        if (seatRepository.existsByScreen_ScreenId(screenId)) {
+            throw new ResourceNotFoundException(
+                    "Cannot delete screen with assigned seats: " + screenId);
+        }
+
         screenRepository.delete(screen);
     }
